Move cluster-arg-resolvers
