+++ src/confluent/tools/handlers/kafka/alter-topic-config.ts
@@ -4,8 +4,8 @@ import {
   CREATE_UPDATE,
   ToolConfig,
 } from "@src/confluent/tools/base-tools.js";
+import { resolveKafkaRestArgs } from "@src/confluent/tools/cluster-arg-resolvers.js";
 import { kafkaRestWithAuthOrOAuth } from "@src/confluent/tools/connection-predicates.js";
-import { resolveKafkaRestArgs } from "@src/confluent/tools/handlers/kafka/cluster-arg-resolvers.js";
 import { ToolName } from "@src/confluent/tools/tool-name.js";
 import { ServerRuntime } from "@src/server-runtime.js";
 import { wrapAsPathBasedClient } from "openapi-fetch";
@@ -33,7 +33,7 @@ const alterTopicConfigArguments = z.object({
     .string()
     .optional()
     .describe(
-      "Confluent Cloud environment ID (env-...) that owns the cluster. Discover via list-environments",
+      "Confluent Cloud environment ID (env-...) that owns the cluster. Discover via list-environments.",
     ),
 });
